examples/trading_terminal/server/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
yfinance
tastytrade
pandas
numpy
